[move]: 통합DB 관련 api를 orig로 경로 및 패키지 변경
