+++ src/main/java/media/wepg/prototype/PrototypeApplication.java
@@ -2,10 +2,8 @@ package media.wepg.prototype;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@EnableJpaRepositories
 public class PrototypeApplication {
 
 	public static void main(String[] args) {
