+++ src/main/java/media/wepg/prototype/orig/repository/ProgramRepository.java
@@ -9,6 +9,7 @@ import java.util.List;
 
 @Repository
 public interface ProgramRepository extends JpaRepository<Program, Long> {
+
     @Query("select pr from Program pr where pr.eventStartDate >= SYSDATE()-1 and pr.eventEndDate <= SYSDATE()+7 and pr.serviceId = :serviceId")
     List<Program> findByServiceId(Long serviceId);
 }
